Implement playout_count for ClusterPlayer.
This involved refactoring accessors related
to playouts and statistics into a separate
interface called StatisticalPlayer.

+++ src/orego/cluster/ClusterPlayerTest.java
@@ -65,9 +65,9 @@ public class ClusterPlayerTest {
 						} catch (InterruptedException e) {
 							e.printStackTrace();
 						}
-						int[] runs = new int[FIRST_POINT_BEYOND_BOARD];
+						long[] runs = new long[FIRST_POINT_BEYOND_BOARD];
 						Arrays.fill(runs, 2);
-						int[] wins = new int[FIRST_POINT_BEYOND_BOARD];
+						long[] wins = new long[FIRST_POINT_BEYOND_BOARD];
 						Arrays.fill(wins, 1);
 						if(bestMove >= 0) {
 							runs[bestMove] += 10;
